docs: ✏️ Fixed stories TS errors and import paths
