+++ src/tasks/components/Layer/index.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { TasksLayerDecorator } from "../../../utils/stories";
+import { TasksLayerDecorator } from "../../../utils/stories/decorators/Tasks";
 
 import TasksLayer from ".";
 
@@ -17,6 +17,8 @@ type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
   args: {
+    onTaskEvent: () => {},
+    setTaskTooltip: () => {},
     taskTooltip: null,
   },
 };
