+++ src/tasks/components/Task/index.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { COLOR_ARG_TYPE, STORY_DATA, TaskDecorator } from "../../../utils/stories";
+import { COLOR_ARG_TYPE, STORY_DATA, TaskDecorator } from "../../../utils/stories/decorators/Tasks";
 
 import { TaskDocs } from ".";
 
@@ -21,6 +21,9 @@ type Story = StoryObj<typeof meta>;
 export const Primary: Story = {
   args: {
     data: STORY_DATA.tasks[0],
+    onLeave: () => {},
+    onOver: () => {},
+    onTaskEvent: () => {},
     width: 100,
     x: 50,
     y: 50,
